Revert "(feat) O3-3773 Support for Patient Program Summary in Patient Chart (…"
This reverts commit 55b4ad5464e73de1b53c4518ee3e506a82915134.
