--- packages/esm-patient-chart-app/src/clinical-views/encounter-tile/tile.scss
@@ -1,87 +1,0 @@
-@use '@carbon/layout';
-@use '@carbon/type';
-@use '@openmrs/esm-styleguide/src/vars' as *;
-
-.title {
-  @include type.type-style('productive-heading-03');
-}
-
-.title:after {
-  content: '';
-  display: block;
-  width: layout.$spacing-07;
-  padding-top: layout.$spacing-01;
-  border-bottom: layout.$spacing-03 solid $brand-teal-01;
-}
-
-.tileTitle {
-  @include type.type-style('body-long-01');
-  display: block;
-  padding-top: layout.$spacing-05;
-}
-
-.columnContainer {
-  display: flex;
-  flex-direction: row;
-}
-
-.tileSubTitle {
-  @include type.type-style('body-compact-01');
-}
-
-.tileValue {
-  @include type.type-style('label-01');
-  margin: layout.$spacing-04 0 layout.$spacing-04;
-  display: block;
-}
-
-.tileBox {
-  display: flex;
-  gap: layout.$spacing-05;
-}
-
-.desktopHeading {
-  h4 {
-    @include type.type-style('heading-compact-02');
-    color: $text-02;
-  }
-}
-
-.tabletHeading {
-  h4 {
-    @include type.type-style('heading-03');
-    color: $text-02;
-  }
-}
-
-.desktopHeading,
-.tabletHeading {
-  text-align: left;
-  text-transform: capitalize;
-  margin-bottom: layout.$spacing-05;
-
-  h4:after {
-    content: '';
-    display: block;
-    width: layout.$spacing-07;
-    padding-top: layout.$spacing-01;
-    border-bottom: 0.375rem solid var(--brand-03);
-  }
-}
-
-.heading:after {
-  content: '';
-  display: block;
-  width: layout.$spacing-07;
-  padding-top: layout.$spacing-01;
-  border-bottom: 0.375rem solid var(--brand-03);
-}
-
-.tile {
-  border: 1px solid $ui-03;
-  margin-bottom: layout.$spacing-03;
-}
-
-[data-extension-id='clinical-views-summary'] {
-  height: auto;
-}
