+++ packages/esm-patient-chart-app/src/encounter-list/components/encounter-list-tabs.scss
