+++ packages/esm-patient-chart-app/src/encounter-list/components/table.scss
