+++ packages/esm-patient-chart-app/src/encounter-list/encounter-list.resource.ts
