+++ packages/esm-patient-chart-app/src/encounter-list/hooks/index.ts
@@ -1,3 +1,2 @@
 export * from './useEncounterRows';
 export * from './useFormsJson';
-export * from './useLastEncounter';
