+++ packages/esm-patient-chart-app/src/encounter-list/hooks/useEncounterRows.ts
