+++ packages/esm-patient-chart-app/src/index.ts
@@ -12,7 +12,7 @@ import genericDashboardComponent, { genericDashboardConfigSchema } from './side-
 import genericNavGroupComponent, { genericNavGroupConfigSchema } from './side-nav/generic-nav-group.component';
 import { moduleName } from './constants';
 import { setupCacheableRoutes, setupOfflineVisitsSync } from './offline';
-import { summaryDashboardMeta, encountersDashboardMeta } from './dashboard.meta';
+import { encountersDashboardMeta, summaryDashboardMeta } from './dashboard.meta';
 import addPastVisitActionButtonComponent from './actions-buttons/add-past-visit.component';
 import cancelVisitActionButtonComponent from './actions-buttons/cancel-visit.component';
 import currentVisitSummaryComponent from './visit/visits-widget/current-visit-summary.component';
@@ -102,11 +102,6 @@ export const stopVisitPatientSearchActionButton = getSyncLifecycle(stopVisitActi
   featureName: 'patient-actions-slot',
   moduleName,
 });
-
-export const clinicalViewsSummary = getAsyncLifecycle(
-  () => import('./clinical-views/encounter-tile/clinical-views-summary.component'),
-  { featureName: 'clinical-views-summary', moduleName },
-);
 
 export const cancelVisitPatientSearchActionButton = getSyncLifecycle(cancelVisitActionButtonComponent, {
   featureName: 'patient-actions-slot-cancel-visit-button',
@@ -232,6 +227,6 @@ export const activeVisitActionsComponent = getAsyncLifecycle(
 );
 
 export const encounterListTableTabs = getAsyncLifecycle(
-  () => import('./clinical-views/encounter-list/encounter-list-tabs.component'),
+  () => import('./encounter-list/components/encounter-list-tabs.component'),
   { featureName: 'encounter-list-table-tabs', moduleName },
 );
