+++ packages/esm-patient-common-lib/src/types/index.ts
@@ -1,5 +1,3 @@
-import { type OpenmrsResource } from '@openmrs/esm-framework';
-
 export * from './test-results';
 
 export interface DashboardLinkConfig {
@@ -56,14 +54,4 @@ export interface HtmlFormEntryForm {
   formUiResource: string;
   formUiPage: 'enterHtmlFormWithSimpleUi' | 'enterHtmlFormWithStandardUi';
   formEditUiPage: 'editHtmlFormWithSimpleUi' | 'editHtmlFormWithStandardUi';
-}
-
-export interface OpenmrsEncounter extends OpenmrsResource {
-  encounterDatetime: Date;
-  encounterType: string;
-  patient: string;
-  location: string;
-  encounterProviders?: Array<{ encounterRole: string; provider: string }>;
-  obs: Array<OpenmrsResource>;
-  form?: string;
 }
